chore: create a model for the current workout
