+++ lib/features/tracker/controllers/current_workout_controller.dart
@@ -1,7 +1,7 @@
 import 'package:flex_workout_mobile/features/exercise/data/models/exercise_model.dart';
+import 'package:flex_workout_mobile/features/tracker/data/models/current_workout_model.dart';
 import 'package:flex_workout_mobile/features/tracker/data/models/tracked_workout_model.dart';
 import 'package:flex_workout_mobile/features/tracker/data/models/tracker_form_model.dart';
-import 'package:flex_workout_mobile/features/tracker/data/models/workout_section_model.dart';
 import 'package:fpdart/fpdart.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
@@ -10,7 +10,7 @@ part 'current_workout_controller.g.dart';
 @Riverpod(keepAlive: true)
 class CurrentWorkoutController extends _$CurrentWorkoutController {
   @override
-  TrackedWorkoutModel build() {
+  CurrentWorkout build() {
     final now = DateTime.now();
     final time = now.hour < 11
         ? 'Morning'
@@ -18,14 +18,10 @@ class CurrentWorkoutController extends _$CurrentWorkoutController {
             ? 'Afternoon'
             : 'Evening';
 
-    return TrackedWorkoutModel(
-      id: -1,
+    return CurrentWorkout(
       title: 'Temp Workout',
       subtitle: '$time Workout',
-      durationInMinutes: -1,
       startTimestamp: now,
-      createdAt: DateTime.now(),
-      updatedAt: DateTime.now(),
     );
   }
 
@@ -34,24 +30,22 @@ class CurrentWorkoutController extends _$CurrentWorkoutController {
     // _updateMuscleGroups(exercises);
 
     final setTypes = exercises.mapWithIndex(
-      (exercise, index) => SetTypeModel(
-        id: 0,
+      (exercise, index) => CurrentWorkoutSetType(
         setLetter: String.fromCharCode(65 + index),
         type: SetTypeEnum.normalSet,
         exercise: exercise,
       ),
     );
 
-    final organizer = SetOrganizerModel(
-      id: 0,
+    final organizer = CurrentWorkoutOrganizer(
       setNumber: 1,
       organization: SetOrganizationEnum.superSet,
       superSet: setTypes.toList(),
     );
 
-    final section = WorkoutSectionModel(
-      id: 0,
+    final section = CurrentWorkoutSection(
       title: _getSuperSetName(exercises),
+      templateOrganizer: organizer,
       organizers: [organizer],
     );
 
@@ -64,22 +58,20 @@ class CurrentWorkoutController extends _$CurrentWorkoutController {
       // _updateMuscleGroups(exercises);
 
       /// Add Set
-      final setType = SetTypeModel(
-        id: 0,
+      final setType = CurrentWorkoutSetType(
         type: SetTypeEnum.normalSet,
         exercise: exercise,
       );
 
-      final organizer = SetOrganizerModel(
-        id: 0,
+      final organizer = CurrentWorkoutOrganizer(
         setNumber: 1,
         organization: SetOrganizationEnum.defaultSet,
         defaultSet: setType,
       );
 
-      final section = WorkoutSectionModel(
-        id: 0,
+      final section = CurrentWorkoutSection(
         title: exercise.name,
+        templateOrganizer: organizer,
         organizers: [organizer],
       );
