+++ lib/features/tracker/data/db/tracked_workout_entity.dart
@@ -1,4 +1,4 @@
-import 'package:flex_workout_mobile/features/tracker/data/db/workout_section_entity.dart';
+import 'package:flex_workout_mobile/features/exercise/data/db/exercise_entity.dart';
 import 'package:flex_workout_mobile/features/tracker/data/models/tracked_workout_model.dart';
 import 'package:objectbox/objectbox.dart';
 
@@ -48,5 +48,113 @@ extension ConvertTrackedWorkout on TrackedWorkout {
         sections: sections.map((e) => e.toModel()).toList(),
         updatedAt: updatedAt,
         createdAt: createdAt,
+      );
+}
+
+@Entity()
+class WorkoutSection {
+  WorkoutSection({
+    required this.title,
+    this.id = 0,
+  });
+
+  @Id()
+  int id = 0;
+
+  String title;
+
+  final workout = ToOne<TrackedWorkout>();
+
+  @Backlink('section')
+  final organizers = ToMany<SetOrganizer>();
+}
+
+extension ConvertWorkoutSection on WorkoutSection {
+  WorkoutSectionModel toModel() => WorkoutSectionModel(
+        id: id,
+        title: title,
+        organizers: organizers.map((e) => e.toModel()).toList(),
+      );
+}
+
+@Entity()
+class SetOrganizer {
+  SetOrganizer({
+    this.id = 0,
+  });
+
+  @Id()
+  int id = 0;
+
+  final section = ToOne<WorkoutSection>();
+
+  @Backlink('superOrganizer')
+  final superSet = ToMany<SetType>();
+
+  final defaultSet = ToOne<SetType>();
+}
+
+extension ConvertSetOrganizer on SetOrganizer {
+  SetOrganizerModel toModel() {
+    SetOrganizationEnum organization;
+
+    if (defaultSet.target != null) {
+      organization = SetOrganizationEnum.defaultSet;
+    } else {
+      organization = SetOrganizationEnum.superSet;
+    }
+
+    return SetOrganizerModel(
+      id: id,
+      organization: organization,
+      setNumber: 0,
+      defaultSet: defaultSet.target?.toModel(),
+      superSet: superSet.map((e) => e.toModel()).toList(),
+    );
+  }
+}
+
+@Entity()
+class SetType {
+  SetType({
+    this.id = 0,
+  });
+
+  @Id()
+  int id = 0;
+
+  final exercise = ToOne<Exercise>();
+
+  final superOrganizer = ToOne<SetOrganizer>();
+
+  final normalSet = ToOne<NormalSet>();
+}
+
+extension ConvertSetType on SetType {
+  SetTypeModel toModel() {
+    const type = SetTypeEnum.normalSet;
+
+    return SetTypeModel(
+      id: id,
+      type: type,
+      exercise: exercise.target!.toModel(),
+      normalSet: normalSet.target?.toModel(),
+    );
+  }
+}
+
+@Entity()
+class NormalSet {
+  NormalSet({
+    this.id = 0,
+  });
+
+  @Id()
+  int id = 0;
+}
+
+extension ConvertNormalSet on NormalSet {
+  NormalSetModel toModel() => NormalSetModel(
+        id: id,
       );
 }
