+++ lib/features/tracker/data/models/tracked_workout_model.dart
@@ -1,9 +1,18 @@
+import 'package:flex_workout_mobile/features/exercise/data/models/exercise_model.dart';
 import 'package:flex_workout_mobile/features/tracker/data/db/tracked_workout_entity.dart';
-import 'package:flex_workout_mobile/features/tracker/data/models/workout_section_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:intl/intl.dart';
 
 part 'tracked_workout_model.freezed.dart';
+
+enum SetOrganizationEnum {
+  defaultSet,
+  superSet,
+}
+
+enum SetTypeEnum {
+  normalSet,
+}
 
 @freezed
 class TrackedWorkoutModel with _$TrackedWorkoutModel {
@@ -31,6 +40,69 @@ extension ConvertTrackedWorkoutModel on TrackedWorkoutModel {
         updatedAt: updatedAt,
         createdAt: createdAt,
       );
+}
+
+@freezed
+class WorkoutSectionModel with _$WorkoutSectionModel {
+  const factory WorkoutSectionModel({
+    required int id,
+    required String title,
+    @Default([]) List<SetOrganizerModel> organizers,
+  }) = _WorkoutSectionModel;
+}
+
+extension ConvertWorkoutSectionModel on WorkoutSectionModel {
+  WorkoutSection toEntity() => WorkoutSection(
+        title: title,
+      )..organizers.addAll(organizers.map((e) => e.toEntity()).toList());
+}
+
+@freezed
+class SetOrganizerModel with _$SetOrganizerModel {
+  const factory SetOrganizerModel({
+    required int id,
+    required SetOrganizationEnum organization,
+    required int setNumber,
+    SetTypeModel? defaultSet,
+    @Default([]) List<SetTypeModel> superSet,
+  }) = _SetOrganizerModel;
+}
+
+extension ConvertSetOrganizerModel on SetOrganizerModel {
+  SetOrganizer toEntity() => SetOrganizer()
+    ..defaultSet.target = organization == SetOrganizationEnum.defaultSet
+        ? defaultSet?.toEntity()
+        : null
+    ..superSet.addAll(superSet.map((e) => e.toEntity()).toList());
+}
+
+@freezed
+class SetTypeModel with _$SetTypeModel {
+  const factory SetTypeModel({
+    required int id,
+    required ExerciseModel exercise,
+    required SetTypeEnum type,
+    String? setLetter,
+    NormalSetModel? normalSet,
+  }) = _SetTypeModel;
+}
+
+extension ConvertSetTypeModel on SetTypeModel {
+  SetType toEntity() => SetType()
+    ..exercise.target = exercise.toEntity()
+    ..normalSet.target =
+        type == SetTypeEnum.normalSet ? normalSet?.toEntity() : null;
+}
+
+@freezed
+class NormalSetModel with _$NormalSetModel {
+  const factory NormalSetModel({
+    required int id,
+  }) = _NormalSetModel;
+}
+
+extension ConvertNormalSetModel on NormalSetModel {
+  NormalSet toEntity() => NormalSet();
 }
 
 extension Date on DateTime {
