+++ lib/features/tracker/ui/containers/set_tiles/default_set_tile.dart
@@ -2,7 +2,7 @@ import 'package:flex_workout_mobile/core/common/ui/components/button.dart';
 import 'package:flex_workout_mobile/core/common/ui/components/flex_list_tile.dart';
 import 'package:flex_workout_mobile/core/extensions/ui_extensions.dart';
 import 'package:flex_workout_mobile/core/theme/app_layout.dart';
-import 'package:flex_workout_mobile/features/tracker/data/models/workout_section_model.dart';
+import 'package:flex_workout_mobile/features/tracker/data/models/current_workout_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_swipe_action_cell/flutter_swipe_action_cell.dart';
@@ -15,7 +15,7 @@ class DefaultSetTile extends ConsumerWidget {
     super.key,
   });
 
-  final SetOrganizerModel organizer;
+  final CurrentWorkoutOrganizer organizer;
   final void Function(VoidCallback) setState;
 
   @override
